+++ firmware/heater_control.h
@@ -6,4 +6,4 @@ void StartHeaterControl();
 bool IsRunningClosedLoop();
 void SetBatteryVoltage(float vbatt);
 void SetHeaterAllowed(bool allowed);
-uint16_t GetHeaterDuty();
+float GetHeaterDuty();
